backend/requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
python-dotenv==1.0.1
google-generativeai==0.3.2
pydantic==2.7.1
sentence-transformers==2.7.0
scikit-learn==1.4.2
torch>=1.13.1
transformers>=4.36.2
cors